(refactor) Move the socket functionality to a new folder
Move the socket pathways to a new folder.
Extract some utility funcitons to a different file as well.
